refactor(capacitor): keep guarded image helper internal
