+++ plugins/plugin-capacitor-bridge/src/mobile-device-bridge-bootstrap.ts
@@ -32,7 +32,6 @@ import {
 	type AgentRuntime,
 	applyBackgroundInferenceBudget,
 	ElizaError,
-	fetchRemoteMedia,
 	type GenerateTextParams,
 	getInferencePriorityGate,
 	type IAgentRuntime,
@@ -48,6 +47,7 @@ import {
 	ServiceType,
 	type TextEmbeddingParams,
 } from "@elizaos/core";
+import { imageUrlToBase64 } from "./image-url-to-base64.ts";
 import { resolveStoredModelPath } from "./shared/local-inference-stored-path.ts";
 
 const DEVICE_BRIDGE_PATH = "/api/local-inference/device-bridge";
@@ -57,10 +57,6 @@ const DEFAULT_NATIVE_REQUEST_TIMEOUT_MS = 600_000;
 const DEFAULT_CALL_TIMEOUT_MS = DEFAULT_NATIVE_REQUEST_TIMEOUT_MS;
 const DEFAULT_LOAD_TIMEOUT_MS = DEFAULT_NATIVE_REQUEST_TIMEOUT_MS;
 const SERVICE_ENABLED = process.env.ELIZA_DEVICE_BRIDGE_ENABLED?.trim() === "1";
-// Bounds for the bionic IMAGE_DESCRIPTION image fetch — same shape the
-// plugin-local-inference vision handlers apply (VISION_IMAGE_* there).
-const IMAGE_DESCRIPTION_FETCH_MAX_BYTES = 20 * 1024 * 1024;
-const IMAGE_DESCRIPTION_FETCH_TIMEOUT_MS = 15_000;
 
 /**
  * Constant-time pairing-token comparison (W1-011). Callers fail closed when
@@ -2181,34 +2177,6 @@ export async function attachMobileDeviceBridgeToServer(
 	server: HttpServer,
 ): Promise<void> {
 	await mobileDeviceBridge.attachToHttpServer(server);
-}
-
-/** Resolve a data:/http(s) image URL to base64 image bytes for the host. */
-export async function imageUrlToBase64(url: string): Promise<string> {
-	if (url.startsWith("data:")) {
-		const comma = url.indexOf(",");
-		return comma >= 0 ? url.slice(comma + 1) : url;
-	}
-	try {
-		// Caller-influenced image URLs must go through the shared SSRF media
-		// guard; a bare `fetch` would let a bionic-delegated phone describe
-		// internal hosts or pull an unbounded payload into memory.
-		const media = await fetchRemoteMedia({
-			url,
-			maxBytes: IMAGE_DESCRIPTION_FETCH_MAX_BYTES,
-			timeoutMs: IMAGE_DESCRIPTION_FETCH_TIMEOUT_MS,
-			maxRedirects: 5,
-		});
-		return media.buffer.toString("base64");
-	} catch (err) {
-		// error-policy:J2 context-adding rethrow — preserve guarded-fetch cause
-		throw new Error(
-			`[mobile-device-bridge] IMAGE_DESCRIPTION failed to fetch ${url}: ${
-				err instanceof Error ? err.message : String(err)
-			}`,
-			{ cause: err },
-		);
-	}
 }
 
 /**
